+++ config.go
@@ -76,6 +76,9 @@ type Config struct {
 	// ping 送信の時間間隔
 	WebSocketPingIntervalSec int32 `ini:"websocket_ping_interval_sec"`
 
+	// シグナリングからコピーする WebSocket の HTTP ヘッダー名
+	CopyWebSocketHeaderNames []string `ini:"copy_websocket_header_names"`
+
 	AuthnWebhookURL      string `ini:"authn_webhook_url"`
 	DisconnectWebhookURL string `ini:"disconnect_webhook_url"`
 
@@ -202,6 +205,8 @@ func (c *Config) PrintConfig() {
 
 	zlog.Info().Str("listen_ipv4_address", c.ListenIPv4Address).Msg("AyameConf")
 	zlog.Info().Int32("listen_port_number", c.ListenPortNumber).Msg("AyameConf")
+
+	zlog.Info().Strs("copy_websocket_header_names", c.CopyWebSocketHeaderNames).Msg("AyameConf")
 
 	zlog.Info().Str("authn_webhook_url", c.AuthnWebhookURL).Msg("AyameConf")
 	zlog.Info().Str("disconnect_webhook_url", c.DisconnectWebhookURL).Msg("AyameConf")