+++ CHANGES.md
@@ -11,6 +11,11 @@
 
 ## develop
 
+- [ADD] シグナリングで利用する WebSocket の HTTP ヘッダーをウェブフックにコピーする `copy_websocket_header_names` 設定を追加する
+  - HTTP ヘッダー名は大文字小文字を区別しません
+  - 複数のヘッダー名を指定できます
+  - @voluntas
+
 ## 2025.3.0
 
 - [CHANGE] シグナリングログファイル名のデフォルトを `signaling.log` から `signaling.jsonl` に変更する